# implement io.StringWriter