--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,3 @@
+module github.com/segmentio/murmur3
+
+go 1.17